requirements.txt
torch>=2.1.0
transformers>=4.35.0
faiss-cpu>=1.7.4
datasets>=2.14.0
openai>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
dacite>=1.8.0
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.14.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
httpx>=0.25.0
